requirements.txt
streamlit
psycopg2-binary
psycopg
pandas
SQLAlchemy
pyodbc==5.1.0
openpyxl
